+++ packages/youtube-comments/client/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './app.css';
-import { App as App2 } from './App2';
+import { App as App2 } from './App';
 
 localStorage.debug = '*,-socket.io-client:socket,-socket.io-parser,-engine.io-client:socket,-engine.io-client:polling'; // For debugging
